requirements.txt
# Core Dependencies
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# AI/LLM Dependencies
openai>=1.0.0

# Excel Processing and Formatting
openpyxl>=3.1.0
xlsxwriter>=3.1.9

# Data Processing
pathlib2>=2.3.7

# Configuration and Logging
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Terminal UI for verification
rich>=13.7.0

# Date and Time Handling
python-dateutil>=2.8.2

# Type Checking and Validation
typing-extensions>=4.8.0

# Error Handling and Retries
tenacity>=8.2.3

# JSON Processing
ujson>=5.8.0

# Visualization (optional)
matplotlib>=3.7.0
plotly>=5.18.0

# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Code Quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# Development Tools
ipython>=8.18.0
jupyter>=1.0.0

# Optional: Fast JSON parsing
orjson>=3.9.0

# Optional: Progress bars
tqdm>=4.66.0